+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Fixed
 - Error when attempting to translate stats with reminder strings. Translation for
   those is currently not supported.
+- An error when translating aliased stats. They were considered as required 
+  params.
 
 ## [0.1.2](https://github.com/eps1lon/poe-mods/compare/v0.1.1...v0.1.2) (2017-20-03)
 ### Added